Restructured styles
