--- src/styles.js
@@ -1,109 +1,0 @@
-const styles = {
-  classicalSquares: {
-    fontFamily: "Arial, Helvetica, sans-serif",
-    display: "grid",
-    gridGap: "0",
-    justifyContent: "center",
-    alignContent: "center",
-    margin: "0 auto"
-  },
-  capablancaSquares: {
-    fontFamily: "Arial, Helvetica, sans-serif",
-    display: "grid",
-    gridGap: "0",
-    justifyContent: "center",
-    alignContent: "center",
-    margin: "0 auto"
-  },
-  userSelect: {
-    WebkitTouchCallout: "none",
-    WebkitUserSelect: "none",
-    KhtmlUserSelect: "none",
-    MozUserSelect: "none",
-    msUserSelect: "none",
-    userSelect: "none",
-  },
-  sq: {
-    position: "relative",
-    textAlign: "center",
-    cursor: "default"
-  },
-  w: {
-    backgroundColor: '#ffce9e',
-  },
-  b: {
-    backgroundColor: '#d18b47',
-  },
-  isLegal: {
-    boxShadow: "inset -4px 0 0 #ffdd43, inset 0 -4px 0 #ffdd43, inset 4px 0 0 #ffeb8e, inset 0 4px 0 #ffeb8e",
-    cursor: "pointer",
-  },
-  isSelected: {
-    backgroundColor: "#ffdd43",
-  },
-  isCheck: {
-    boxShadow: "inset -5px 0 0 #fdd57d, inset 0 -5px 0 #fdd57d, inset 5px 0 0 #fdd57d, inset 0 5px 0 #fdd57d"
-  },
-  file: {
-    position: "absolute",
-    fontWeight: "bold",
-    fontSize: "0.8rem",
-    marginRight: "0.15rem",
-    marginBottom: "0.15rem",
-    bottom: "0",
-    right: "0"
-  },
-  rank: {
-    position: "absolute",
-    fontWeight: "bold",
-    fontSize: "0.8rem",
-    marginLeft: "0.15rem",
-    marginTop: "0.15rem",
-    top: "0",
-    left: "0"
-  },
-  light: {
-    color: '#d18b47',
-  },
-  dark: {
-    color: '#ffce9e',
-  },
-  slider:{
-    width: "100%",
-    background: "#d3d3d3",
-    WebkitTransition: ".2s",
-    transition: "opacity .2s",
-    display: "grid",
-    margin: "15px auto",
-  },
-  movesTable: {
-    table: {
-      width: '100%',
-      height: 368,
-    },
-    tbody: {
-      display: 'table-row-group',
-    },
-    td: {
-      fontWeight: 400,
-      fontSize: "0.875rem",
-      lineHeight: 1.43,
-      display: "table-cell",
-      borderBottom: "1px solid rgb(224, 224, 224)",
-      textAlign: "left",
-      padding: "6px 16px",
-      background: "#ffffff",
-    },
-    current: {
-      color: '#ffffff',
-      background: '#1976d2',
-      fontWeight: 'bold',
-    },
-    tr: {
-      display: 'table-row',
-      verticalAlign: 'middle',
-    },
-  },
-}
-
-export default styles;
