+++ src/styles/algebraicNotation.js
@@ -1,0 +1,28 @@
+const styles = {
+  file: {
+    position: "absolute",
+    fontWeight: "bold",
+    fontSize: "0.8rem",
+    marginRight: "0.15rem",
+    marginBottom: "0.15rem",
+    bottom: "0",
+    right: "0"
+  },
+  rank: {
+    position: "absolute",
+    fontWeight: "bold",
+    fontSize: "0.8rem",
+    marginLeft: "0.15rem",
+    marginTop: "0.15rem",
+    top: "0",
+    left: "0"
+  },
+  light: {
+    color: '#d18b47',
+  },
+  dark: {
+    color: '#ffce9e',
+  },
+}
+
+export default styles;
