+++ src/styles/movesTable.js
@@ -1,0 +1,31 @@
+const styles = {
+  table: {
+    width: '100%',
+    height: 368,
+    overflowY: 'scroll',
+    display: 'block',
+  },
+  tbody: {
+  },
+  td: {
+    fontWeight: 400,
+    fontSize: "0.875rem",
+    lineHeight: 1.43,
+    display: "table-cell",
+    textAlign: "left",
+    padding: "6px 16px",
+    background: "#ffffff",
+  },
+  current: {
+    color: '#ffffff',
+    background: '#1976d2',
+    fontWeight: 'bold',
+  },
+  tr: {
+    display: 'table-row',
+    verticalAlign: 'middle',
+    height: '2.5rem',
+  },
+}
+
+export default styles;
