+++ src/styles/slider.js
@@ -1,0 +1,10 @@
+const styles = {
+  width: "100%",
+  background: "#d3d3d3",
+  WebkitTransition: ".2s",
+  transition: "opacity .2s",
+  display: "grid",
+  margin: "15px auto",
+}
+
+export default styles;
